huggingface/transformers/text2text-generation/pegasus/requirements.txt
optimum-rbln
sentencepiece>=0.2.0
